Minor Fixes

+++ Task/Models/SingleTriangleProperties.cs
@@ -10,17 +10,17 @@ namespace Task.Models
         
 
 
-        public decimal AngularV1x { get; set; }
+        public int AngularV1x { get; set; }
 
-        public decimal AngularV1y { get; set; }
+        public int AngularV1y { get; set; }
 
-        public decimal LeftV2x { get; set; }
+        public int LeftV2x { get; set; }
 
-        public decimal LeftV2y { get; set; }
+        public int LeftV2y { get; set; }
 
-        public decimal RightV3x { get; set; }
+        public int RightV3x { get; set; }
 
-        public decimal RightV3y { get; set; }
+        public int RightV3y { get; set; }
 
 
     }
